requirements.txt
Flask>=2.0
requests>=2.0
python-dotenv>=1.0
gunicorn>=20.0
Flask-SQLAlchemy>=3.0
